+++ module/utility/damage.js
@@ -327,14 +327,15 @@ export class HeroRoll extends Roll {
 export function CombatSkillLevelsForAttack(item) {
     let result = {
         ocv: 0,
+        omcv: 0,
         dcv: 0,
         dmcv: 0,
         omcv: 0,
         dc: 0,
     }
 
-    let csl = item.actor.items.find(o => o.system.XMLID === "COMBAT_LEVELS" && o.system.attacks && o.system.attacks[item.id])
-    if (csl) {
+    let csl = item.actor.items.find(o => ["MENTAL_COMBAT_LEVELS", "COMBAT_LEVELS"].includes(o.system.XMLID) && o.system.attacks && o.system.attacks[item.id])
+    if (csl && csl.system.csl) {
         for (let i = 0; i < parseInt(csl.system.LEVELS.value); i++) {
             result[csl.system.csl[i]] = (result[csl.system.csl[i]] || 0) + 1;
         }
@@ -343,6 +344,6 @@ export function CombatSkillLevelsForAttack(item) {
 
     // Takes 2 CLS for +1 DC
     result.dc = Math.floor(result.dc / 2);
-    
+
     return result;
 }