+++ module/utility/upload_hdc.js
@@ -417,11 +417,17 @@ export async function applyCharacterSheet(xmlDoc) {
 
 
     // Combat Skill Levels - Enumerate attacks that use OCV
-    for (let cslItem of this.actor.items.filter(o => o.system.XMLID === "COMBAT_LEVELS")) {
+    for (let cslItem of this.actor.items.filter(o => ["MENTAL_COMBAT_LEVELS", "COMBAT_LEVELS"].includes(o.system.XMLID))) {
+        let _ocv = 'ocv'
+        if (cslItem.system.XMLID === "MENTAL_COMBAT_LEVELS") {
+            _ocv = 'omcv'
+        }
+
         let attacks = {}
+        let checkedCount = 0;
         for (let attack of this.actor.items.filter(o =>
             (o.type == 'attack' || o.system.subType == 'attack') &&
-            o.system.uses === 'ocv'
+            o.system.uses === _ocv
         )) {
             let checked = false;
 
@@ -452,8 +458,21 @@ export async function applyCharacterSheet(xmlDoc) {
                 checked = true;
             }
 
+            if (cslItem.system.OPTION === "BROAD" && cslItem.system.XMLID === "MENTAL_COMBAT_LEVELS") {
+                checked = true;
+            }
+
             attacks[attack.id] = checked;
+
+            if (checked) checkedCount++;
+
         }
+
+        // Make sure at least one attacked is checked
+        if (checkedCount === 0 && Object.keys(attacks).length > 0) {
+            attacks[Object.keys(attacks)[0]] = true;
+        }
+
         await cslItem.update({ 'system.attacks': attacks });
     }
 
@@ -1269,9 +1288,9 @@ function calcActivePoints(_basePointsPlusAdders, system) {
     // const xmlid = system.rules || system.xmlid //xmlItem.getAttribute('XMLID')
     // const modifiers = system.modifiers || system.MODIFIER || [] //xmlItem.getElementsByTagName("ADDER")
 
-    if (system.XMLID == "MINDCONTROL")
-        HEROSYS.log(false, system.XMLID)
-
+    if (system.XMLID == "MINDCONTROL") {
+        // HEROSYS.log(false, system.XMLID)
+    }
     // NAKEDMODIFIER uses PRIVATE=="Yes" to indicate advantages
 
     let advantages = 0;
@@ -1452,7 +1471,7 @@ export async function uploadPower(power, type) {
     else {
         if (game.settings.get(game.system.id, 'alphaTesting')) {
             ui.notifications.warn(`${xmlid} not handled during HDC upload of ${this.actor.name}`)
-            HEROSYS.log(false, power)
+            // HEROSYS.log(false, power)
         }
 
     }
@@ -2221,6 +2240,13 @@ export async function makeAttack(item) {
         changes[`system.noHitLocations`] = true
     }
 
+    // IMAGES (not implemented)
+    if (xmlid == "IMAGES") {
+        changes[`system.class`] = 'images'
+        changes[`system.usesStrength`] = false
+        changes[`system.noHitLocations`] = true
+    }
+
     // DRAIN (not implemented)
     if (xmlid == "DRAIN") {
         changes[`system.class`] = 'drain'
@@ -2767,20 +2793,23 @@ export async function updateItem(item) {
     if (item.system.LEVELS) {
         if (item.system.LEVELS.value == undefined) {
             let levels = item.system.LEVELS
-            delete item.system.LEVELS
-            item.system.LEVELS = {
+            //delete item.system.LEVELS
+            let _LEVELS = {
                 value: levels,
                 max: levels
             }
             changed = true;
-            await item.update({ 'system.LEVELS.value': levels, 'system.LEVELS.max': levels })
+            //await item.update({ 'system.LEVELS.value': levels, 'system.LEVELS.max': levels })
+            await item.update({ 'system.LEVELS': null })
+            let _item = await item.update({ ['system.LEVELS.value']: levels, ['system.LEVELS.max']: levels})
+            //console.log(_item.system.LEVELS)
         }
 
         // Default values = max
-        if (item.system.LEVELS.value != item.system.LEVELS.max) {
-            item.system.LEVELS.value = parseInt(item.system.LEVELS.max)
-            changed = true
-        }
+        // if (item.system.LEVELS.value != item.system.LEVELS.max) {
+        //     item.system.LEVELS.value = parseInt(item.system.LEVELS.max)
+        //     changed = true
+        // }
 
         // Look for active effects
         if (item.actor.effects) {
@@ -2796,7 +2825,7 @@ export async function updateItem(item) {
                                 item.system.LEVELS.value += parseFloat(change.value / ActivePointsPerLevel) || 0
                                 break;
                             default:
-                                HEROSYS.log(false, "unknown mode")
+                            // HEROSYS.log(false, "unknown mode")
                         }
                     }
                 }
@@ -2853,6 +2882,7 @@ export async function updateItem(item) {
             //         ui.notifications.warn(`${item.actor.name} missing id`)
             //     }
             // } else {
+            
             await item.update({ 'system.description': item.system.description })
             //}
         }